+++ R/bmm_model_mixture2p.R
@@ -76,9 +76,9 @@
 #' fit <- bmm(formula = ff,
 #'            data = dat,
 #'            model = model,
-#'            parallel=T,
-#'            iter=500,
-#'            backend='cmdstanr')
+#'            cores = 4,
+#'            iter = 500,
+#'            backend = 'cmdstanr')
 #' }
 #' @export
 mixture2p <- function(resp_error,