+++ tests/testthat/test-helpers-postprocess.R
@@ -1,15 +1,8 @@
 test_that("bmm version is added to mock model", {
-  dat <- data.frame(y = rsdm(n=10))
-
-  ff <- bmmformula(c ~ 1,
-                   kappa ~ 1)
-
-  fit <- bmm(formula = ff,
-             data = dat,
+  fit <- bmm(formula = bmmformula(c ~ 1, kappa ~ 1),
+             data = data.frame(y = rsdm(n=10)),
              model = sdmSimple(resp_error = "y"),
-             parallel=T,
-             iter=500,
-             backend='mock',
+             backend = 'mock',
              rename = F,
              mock = 1)
   expect_true("bmm" %in% names(fit$version))
@@ -17,7 +10,8 @@ test_that("bmm version is added to mock model", {
 
 
 test_that("get_mu_pars works",{
-  a <- brm(y~ a, data.frame(y = c(1,2,3), a = c('A',"B","C")), backend = "mock", mock_fit = 1, rename=F)
+  a <- brm(y~ a, data.frame(y = c(1,2,3), a = c('A',"B","C")),
+           backend = "mock", mock_fit = 1, rename=F)
   mus <- get_mu_pars(a)
   expect_equal(mus, c("Intercept", "aB", "aC"))
 })