+++ NEWS.md
@@ -20,6 +20,7 @@
 * the function `fit_model()` is deprecated in favor of `bmm()` (#163) and will be removed in a future version
 * the argument **setsize** for the `mixture3p` and `IMM` models is now called **set_size** for consistency (#163). The old argument name is deprecated and will be removed in a future version
 * the distributions functions for the imm model are renamed from dIMM, pIMM, rIMM and qIMM to dimm, pimm, rimm and qimm (#163)
+* the argument parallel for the bmm() function is deprecated and will be removed in a future version. Use `cores` instead, as for brms::brm() (#163)
 
 ### Other changes
 * `bmm` now requires the latest version of `brms` (>= 2.21.0).